+++ bindings/matrix-sdk-ffi/src/client_builder.rs
@@ -183,9 +183,23 @@ impl From<qrcode::LoginProgress> for QrLoginProgress {
 #[uniffi(flat_error)]
 pub enum ClientBuildError {
     #[error(transparent)]
-    Sdk(#[from] MatrixClientBuildError),
+    Sdk(MatrixClientBuildError),
+    #[error("The homeserver doesn't provide a trusted sliding sync proxy in its well-known configuration.")]
+    SlidingSyncNotAvailable,
+
     #[error("Failed to build the client: {message}")]
     Generic { message: String },
+}
+
+impl From<MatrixClientBuildError> for ClientBuildError {
+    fn from(e: MatrixClientBuildError) -> Self {
+        match e {
+            MatrixClientBuildError::SlidingSyncNotAvailable => {
+                ClientBuildError::SlidingSyncNotAvailable
+            }
+            _ => ClientBuildError::Sdk(e),
+        }
+    }
 }
 
 impl From<IdParseError> for ClientBuildError {
@@ -220,6 +234,7 @@ pub struct ClientBuilder {
     server_versions: Option<Vec<String>>,
     passphrase: Zeroizing<Option<String>>,
     user_agent: Option<String>,
+    requires_sliding_sync: bool,
     sliding_sync_proxy: Option<String>,
     proxy: Option<String>,
     disable_ssl_verification: bool,
@@ -242,6 +257,7 @@ impl ClientBuilder {
             server_versions: None,
             passphrase: Zeroizing::new(None),
             user_agent: None,
+            requires_sliding_sync: false,
             sliding_sync_proxy: None,
             proxy: None,
             disable_ssl_verification: false,
@@ -326,6 +342,12 @@ impl ClientBuilder {
     pub fn user_agent(self: Arc<Self>, user_agent: String) -> Arc<Self> {
         let mut builder = unwrap_or_clone_arc(self);
         builder.user_agent = Some(user_agent);
+        Arc::new(builder)
+    }
+
+    pub fn requires_sliding_sync(self: Arc<Self>) -> Arc<Self> {
+        let mut builder = unwrap_or_clone_arc(self);
+        builder.requires_sliding_sync = true;
         Arc::new(builder)
     }
 
@@ -417,14 +439,15 @@ impl ClientBuilder {
         if let QrCodeModeData::Reciprocate { server_name } = &qr_code_data.inner.mode_data {
             let builder = self.server_name_or_homeserver_url(server_name.to_owned());
 
-            let client = builder.build().await.map_err(|e| {
-                error!("Couldn't build the client {e:?}");
-                HumanQrLoginError::Unknown
+            let client = builder.build().await.map_err(|e| match e {
+                ClientBuildError::SlidingSyncNotAvailable => {
+                    HumanQrLoginError::SlidingSyncNotAvailable
+                }
+                _ => {
+                    error!("Couldn't build the client {e:?}");
+                    HumanQrLoginError::Unknown
+                }
             })?;
-
-            if client.sliding_sync_proxy().is_none() {
-                return Err(HumanQrLoginError::SlidingSyncNotAvailable);
-            }
 
             let client_metadata = oidc_configuration
                 .try_into()
@@ -558,6 +581,10 @@ impl ClientBuilder {
         }
 
         inner_builder = inner_builder.with_encryption_settings(builder.encryption_settings);
+
+        if builder.requires_sliding_sync {
+            inner_builder = inner_builder.requires_sliding_sync();
+        }
 
         let sdk_client = inner_builder.build().await?;