+++ crates/matrix-sdk-crypto/src/machine.rs
@@ -867,6 +867,7 @@ impl OlmMachine {
     }
 
     fn add_withheld_info(&self, changes: &mut Changes, event: &RoomKeyWithheldEvent) {
+        debug!(?event.content, "Processing `m.room_key.withheld` event");
         if let RoomKeyWithheldContent::MegolmV1AesSha2(
             MegolmV1AesSha2WithheldContent::BlackListed(c)
             | MegolmV1AesSha2WithheldContent::Unverified(c),
@@ -1190,7 +1191,7 @@ impl OlmMachine {
             }
         };
 
-        debug!(?event, "Received a to-device event");
+        debug!("Received a to-device event");
 
         match event {
             ToDeviceEvents::RoomEncrypted(e) => {