+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Changes:
 
+- Log the content of received `m.room_key.withheld` to-device events.
+  ([#3591](https://github.com/matrix-org/matrix-rust-sdk/pull/3591))
+
 - Attempt to decrypt bundled events (reactions and the latest thread reply) if
   they are found in the unsigned part of an event.
   ([#3468](https://github.com/matrix-org/matrix-rust-sdk/pull/3468))
@@ -108,7 +111,7 @@ Additions:
   ([#3194](https://github.com/matrix-org/matrix-rust-sdk/pull/3194))
 
 
-## 0.7.1 
+## 0.7.1
 
 Security fixes: